Update MLE fit parameter initialisation
Updated the initialisation of the start parameters to use the values from the input configuration. The result is the same as the hardcoded "10^(-1 + 2*rand())" but can be changed via configuration.

The initialisation of the 2nd and 3rd species for konstart must be expressed relative to koffstart (as is done for species 1) and not absolute (i.e. multiplied by koffstart).

This change also adds the best nll score to the returned values as this is of use when selecting fits with different settings, e.g. number of species or fixed parameters.