Update switching probabilities in the simulation to use exponential CDF
When the Steptime is 1e-5 the CDF is approximately the same as the scaled rate constant:

kt ~ 1-e^-kt

This approximation is correct to 5e-4 relative error when rate is 50 and gets worse as the rate increases. The error approaches 1% when rate is around 2000. The approximation results in a probability that is too high and so the switching will be greater than expected.

Use of the exponential CDF ensures support for larger values of the rate constant.